exclude or fix some MiMa issues restructure tests, to allow for additional test to apply to both implementations of BitSet (actual tests to follow)
