+++ build.sbt
@@ -343,6 +343,22 @@ val mimaFilterSettings = Seq {
     ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.immutable.HashMap#HashTrieMap.size0_="),
     ProblemFilters.exclude[FinalMethodProblem]("scala.collection.immutable.HashMap#HashTrieMap.size"),
 
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.BitSetLike.lastWordSet"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.immutable.BitSet.lastWordSet"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.mutable.BitSet.lastWordSet"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.BitSetLike._word"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.immutable.BitSet._word"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.mutable.BitSet._word"),
+    ProblemFilters.exclude[ReversedMissingMethodProblem]("scala.collection.BitSetLike.scala$collection$BitSetLike$$super$intersect"),
+    ProblemFilters.exclude[ReversedMissingMethodProblem]("scala.collection.BitSetLike.scala$collection$BitSetLike$$super$union"),
+    ProblemFilters.exclude[ReversedMissingMethodProblem]("scala.collection.BitSetLike.scala$collection$BitSetLike$$super$diff"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.BitSetLike.filterImpl"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.immutable.BitSet.filterImpl"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.mutable.BitSet.filterImpl"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.BitSetLike.copyOrSelf"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.immutable.BitSet.copyOrSelf"),
+    ProblemFilters.exclude[DirectMissingMethodProblem]("scala.collection.mutable.BitSet.copyOrSelf"),
+
     ProblemFilters.exclude[MissingClassProblem]("scala.collection.immutable.Map$MapNIterator"),
     ProblemFilters.exclude[MissingClassProblem]("scala.collection.immutable.Map$Map2$Map2Iterator"),
     ProblemFilters.exclude[MissingClassProblem]("scala.collection.immutable.Map$Map3$Map3Iterator"),
